improved inherit and cleaned code
