+++ Game/Architecture/Background.cs
@@ -5,23 +5,12 @@ using System.Reflection.Emit;
 using System.Windows.Forms;
 
 namespace Game {
-    public class Background: PictureBox {
-        public Background(Size size){
-            Location = new Point(0,0);
-            Size = size;
+    public class Background: StaticObject {
+        public Background(Size size, Point location)
+        :base( size, location){
             SizeMode = PictureBoxSizeMode.StretchImage;
             Tag = "background";
             Image = new Bitmap(PathToImages + "background.png");
-            Visible = true;
-        }
-        protected string PathToImages = GetGameDirectoryRoot().FullName.ToString() + "\\Images\\";
-        private static DirectoryInfo GetGameDirectoryRoot() {
-            var dir = new DirectoryInfo(Directory.GetCurrentDirectory());
-            while (!dir.ToString().EndsWith("GameOfTheCentury")) {
-                dir = dir.Parent;
-            }
-
-            return dir;
         }
     }
 }
