+++ Game/Architecture/Plant.cs
@@ -12,7 +12,7 @@ namespace Game {
         secondTree,
         fir,
     }
-    public class Plant : PictureBox {
+    public class Plant : StaticObject {
         private Dictionary<PlantsType, string> plantsType = new Dictionary<PlantsType, string>() {
             {PlantsType.firstBush, "firstBush.png"},
             {PlantsType.secodBush, "secodBush.png"},
@@ -20,22 +20,10 @@ namespace Game {
             {PlantsType.secondTree, "secondTree.png"},
             {PlantsType.fir, "thirdTree.png"},
         };
-        public Plant(Point location, Size size, PlantsType plantType){
-            Location = location;
-            Size = size;
-            SizeMode = PictureBoxSizeMode.StretchImage;
+        public Plant(Point location, Size size, PlantsType plantType)
+            :base( size, location){
             Tag = "plant";
             Image = new Bitmap(PathToImages + plantsType[plantType]);
-            Visible = true;
-        }
-        protected string PathToImages = GetGameDirectoryRoot().FullName.ToString() + "\\Images\\";
-        private static DirectoryInfo GetGameDirectoryRoot() {
-            var dir = new DirectoryInfo(Directory.GetCurrentDirectory());
-            while (!dir.ToString().EndsWith("GameOfTheCentury")) {
-                dir = dir.Parent;
-            }
-
-            return dir;
         }
     }
 }
