+++ Game/Architecture/StaticObject.cs
@@ -4,50 +4,30 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Drawing;
+using System.IO;
+using System.Windows.Forms;
 
 namespace Game
 {
-    public class StaticObject : Entity
-    {
+    public class StaticObject : PictureBox {
         public StaticObject() { }
-        public StaticObject(string name, string imageName, int drawingPriority, 
-            Point location, Size size) 
-            : base(name, imageName, drawingPriority, location, size)
-        {
-        }
-
-        public StaticObject WithName(string name)
-        {
-            Name = name;
-            return this;
-        }
-
-        public StaticObject WithImageName(string name)
-        {
-            ImageName = name;
-            return this;
-        }
-
-        public StaticObject WithDrowingPriority(int drawingPriority)
-        {
-            DrawingPriority = drawingPriority;
-            return this;
-        }
-
-        public StaticObject WithLocation(Point location)
-        {
+        
+        public StaticObject(Size size, Point location){
             Location = location;
-            return this;
-        }
-
-        public StaticObject WithSize(Size size)
-        {
             Size = size;
-            return this;
+            SizeMode = PictureBoxSizeMode.StretchImage;
+            //Tag = "background";
+            //Image = new Bitmap(PathToImages + "background.png");
+            Visible = true;
         }
-
-        public override void OnConflict(Entity entity)
-        {
+        
+        protected string PathToImages = GetGameDirectoryRoot().FullName.ToString() + "\\Images\\";
+        private static DirectoryInfo GetGameDirectoryRoot() {
+            var dir = new DirectoryInfo(Directory.GetCurrentDirectory());
+            while (!dir.ToString().EndsWith("GameOfTheCentury")) {
+                dir = dir.Parent;
+            }
+            return dir;
         }
     }
 }
