+++ Game/Architecture/StaticObjectBuilder.cs
@@ -10,59 +10,59 @@ using System.Windows.Forms;
 
 namespace Game
 {
-    public class StaticObjectBuilder : IStaticObjectBuilderInterface
+    public class StaticObjectBuilder
     {
-        public string PathToImages = new DirectoryInfo(Directory.GetCurrentDirectory())
-            .Parent.Parent.Parent.Parent.FullName + "\\Images\\";
-        public StaticObject StaticObject = new StaticObject();
-        public PictureBox PictureBox = new PictureBox();
-        public StaticObjectBuilder() 
-        { 
-            PictureBox.SizeMode = PictureBoxSizeMode.AutoSize;
-            PictureBox.Visible = true;
-        }
-        public Entity BuildEntity()
-        {
-            return StaticObject;
-        }
-
-        public PictureBox BuildPictureBox()
-        {
-            return PictureBox;
-        }
-
-        public StaticObjectBuilder SetName(string name)
-        {
-            StaticObject = StaticObject.WithName(name);
-            PictureBox.Tag = name;
-            return this;
-        }
-
-        public StaticObjectBuilder SetImageName(string imageName)
-        {
-            StaticObject = StaticObject.WithImageName(imageName);
-            PictureBox.BackgroundImage = Image.FromFile(PathToImages + imageName);
-            return this;
-        }
-
-        public StaticObjectBuilder SetDrowingPriority(int drawingPriority)
-        {
-            StaticObject = StaticObject.WithDrowingPriority(drawingPriority);
-            return this;
-        }
-
-        public StaticObjectBuilder SetLocation(Point location)
-        {
-            StaticObject = StaticObject.WithLocation(location);
-            PictureBox.Location = location;
-            return this;
-        }
-
-        public StaticObjectBuilder SetSize(Size size)
-        {
-            StaticObject = StaticObject.WithSize(size);
-            PictureBox.Size = size;
-            return this;
-        }
+        //public string PathToImages = new DirectoryInfo(Directory.GetCurrentDirectory())
+        //    .Parent.Parent.Parent.Parent.FullName + "\\Images\\";
+        //public StaticObject StaticObject = new StaticObject();
+        //public PictureBox PictureBox = new PictureBox();
+        //public StaticObjectBuilder() 
+        //{ 
+        //    PictureBox.SizeMode = PictureBoxSizeMode.AutoSize;
+        //    PictureBox.Visible = true;
+        //}
+        //public Entity BuildEntity()
+        //{
+        //    return StaticObject;
+        //}
+//
+        //public PictureBox BuildPictureBox()
+        //{
+        //    return PictureBox;
+        //}
+//
+        //public StaticObjectBuilder SetName(string name)
+        //{
+        //    StaticObject = StaticObject.WithName(name);
+        //    PictureBox.Tag = name;
+        //    return this;
+        //}
+//
+        //public StaticObjectBuilder SetImageName(string imageName)
+        //{
+        //    StaticObject = StaticObject.WithImageName(imageName);
+        //    PictureBox.BackgroundImage = Image.FromFile(PathToImages + imageName);
+        //    return this;
+        //}
+//
+        //public StaticObjectBuilder SetDrowingPriority(int drawingPriority)
+        //{
+        //    StaticObject = StaticObject.WithDrowingPriority(drawingPriority);
+        //    return this;
+        //}
+//
+        //public StaticObjectBuilder SetLocation(Point location)
+        //{
+        //    StaticObject = StaticObject.WithLocation(location);
+        //    PictureBox.Location = location;
+        //    return this;
+        //}
+//
+        //public StaticObjectBuilder SetSize(Size size)
+        //{
+        //    StaticObject = StaticObject.WithSize(size);
+        //    PictureBox.Size = size;
+        //    return this;
+        //}
     }
 }
